+++ src/Cqrs/Queue/ExecuteQueue.php
@@ -11,10 +11,10 @@ use Xakki\Emailer\Repository;
 
 class ExecuteQueue extends AbstractQueue
 {
-    public function __construct(Emailer $emailer, array $skipIds = [], array $skipProjectIds = [])
+    public function __construct(Emailer $emailer, array $skipIds = [], array $skipTransportIds = [])
     {
         $this->emailer = $emailer;
-        $row = Repository\Queue::findOneByStatus(Queue::QUEUE_STATUS_NEW, true, $skipIds, $skipProjectIds);
+        $row = Repository\Queue::findOneByStatus(Queue::QUEUE_STATUS_NEW, true, $skipIds, $skipTransportIds);
         if (!$row) {
             $e = new DataNotFound('Not found data');
             $e->httpCode = 0;